Fix freeup bug and loadexample2 branch error
Whenever we remove a functional unit element, we need to change
it's computed state to false. If we don't change it, the result
of previoius FU is propagated to the ROB, which we don't want.

This commit also make sure that Loading Example2 from program
builder works properly. Before break used to have drop down menu
instead of text. This fixes that.

Signed-off-by: Mohana Datta Yelugoti <ymdatta.work@gmail.com>